fetch/auth

+++ services/contactsServices.js
@@ -6,16 +6,16 @@ import { v4 as uuidv4 } from 'uuid';
 import {Contact} from "../models/contactModels.js";
 
 
-export const listContacts = () => Contact.find();
+export const listContacts = (ownerId) => Contact.find(owner);
 
-export const getContactById = (id) => Contact.findById(id);
+export const getContactById = (id, ownerId) => Contact.findById({ _id: id, ownerId });
 
-export const removeContact = (id) => Contact.findByIdAndDelete(id);
+export const removeContact = (id, ownerId) => Contact.findByIdAndDelete({ _id: id, ownerId });
 
 
-export const addContact = async (value) => {
+export const addContact = async (value, ownerId) => {
     try {
-        const newContact = value.favorite ? value : { ...value, "favorite": "false" }
+        const newContact = value.favorite ? { ...value, ownerId } : { ...value, "favorite": "false", ownerId }
 
         const addContact = await Contact.create(newContact);
         return addContact; 
@@ -25,10 +25,10 @@ export const addContact = async (value) => {
     
 }
 
-export const contactsService = async (id, data) => {
+export const contactsService = async (id, data, ownerId) => {
     try {
 
-        const updateContact = await Contact.findByIdAndUpdate(id, data, { new: true });
+        const updateContact = await Contact.findByIdAndUpdate(id, data, { new: true }).where("ownerId").equals(ownerId);
         console.log(updateContact);
 
         return updateContact;
@@ -37,9 +37,9 @@ export const contactsService = async (id, data) => {
     }
 }
 
-export const updateStatusFavorite = async (id, data) => {
+export const updateStatusFavorite = async (id, data, ownerId) => {
     try {
-        const test = await Contact.findById(id);
+        const test = await Contact.findById(id, ownerId);
         console.log(test);
         const statusContact = await Contact.findByIdAndUpdate(id, {favorite: data}, {new: true});
         console.log(statusContact);
